enables standard tests to work with filesystem and pyiceberg destinations
