+++ tests/load/pipeline/test_drop.py
@@ -18,7 +18,7 @@ from dlt.pipeline.exceptions import (
 )
 from dlt.destinations.job_client_impl import SqlJobClientBase
 
-from tests.load.utils import destinations_configs, DestinationTestConfiguration
+from tests.load.utils import FILE_BUCKET, destinations_configs, DestinationTestConfiguration
 from tests.pipeline.utils import assert_load_info, load_table_counts
 
 
@@ -143,7 +143,10 @@ def assert_destination_state_loaded(pipeline: Pipeline) -> None:
 @pytest.mark.parametrize(
     "destination_config",
     destinations_configs(
-        default_sql_configs=True, local_filesystem_configs=True, all_buckets_filesystem_configs=True
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        all_buckets_filesystem_configs=True,
+        table_format_filesystem_configs=True,
     ),
     ids=lambda x: x.name,
 )
@@ -232,7 +235,12 @@ def test_drop_command_resources_and_state(
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_drop_command_only_state(destination_config: DestinationTestConfiguration) -> None:
@@ -259,7 +267,12 @@ def test_drop_command_only_state(destination_config: DestinationTestConfiguratio
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_drop_command_only_tables(destination_config: DestinationTestConfiguration) -> None:
@@ -283,7 +296,12 @@ def test_drop_command_only_tables(destination_config: DestinationTestConfigurati
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_drop_destination_tables_fails(destination_config: DestinationTestConfiguration) -> None:
@@ -313,7 +331,12 @@ def test_drop_destination_tables_fails(destination_config: DestinationTestConfig
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_fail_after_drop_tables(destination_config: DestinationTestConfiguration) -> None:
@@ -346,7 +369,12 @@ def test_fail_after_drop_tables(destination_config: DestinationTestConfiguration
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_load_step_fails(destination_config: DestinationTestConfiguration) -> None:
@@ -391,7 +419,12 @@ def test_resource_regex(destination_config: DestinationTestConfiguration) -> Non
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_drop_nothing(destination_config: DestinationTestConfiguration) -> None:
@@ -410,7 +443,11 @@ def test_drop_nothing(destination_config: DestinationTestConfiguration) -> None:
 
 
 @pytest.mark.parametrize(
-    "destination_config", destinations_configs(default_sql_configs=True), ids=lambda x: x.name
+    "destination_config",
+    destinations_configs(
+        default_sql_configs=True, table_format_filesystem_configs=True, bucket_subset=(FILE_BUCKET,)
+    ),
+    ids=lambda x: x.name,
 )
 def test_drop_all_flag(destination_config: DestinationTestConfiguration) -> None:
     """Using drop_all flag. Destination dataset and all local state is deleted"""
@@ -437,7 +474,12 @@ def test_drop_all_flag(destination_config: DestinationTestConfiguration) -> None
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_run_pipeline_after_partial_drop(destination_config: DestinationTestConfiguration) -> None:
@@ -458,7 +500,12 @@ def test_run_pipeline_after_partial_drop(destination_config: DestinationTestConf
 
 @pytest.mark.parametrize(
     "destination_config",
-    destinations_configs(default_sql_configs=True, local_filesystem_configs=True),
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        bucket_subset=(FILE_BUCKET,),
+    ),
     ids=lambda x: x.name,
 )
 def test_drop_state_only(destination_config: DestinationTestConfiguration) -> None:
